form_confirmed added

+++ basic/views/form/fill-form.php
@@ -28,11 +28,17 @@ foreach($blocks as $block){
                         $val = $submitteds[$j]->value;
                     }
                     $submitteds[$j]->value = $val;
-                    echo $form->field($submitteds[$j], "[$j]value")->label('Балл');
-                    echo  $form->field($submitteds[$j], 'criteria_id')->hiddenInput(['value'=> $criterias[$i]->id])->label(false);
-                    echo '<input type="hidden" name="id" value="'.$submitteds[$j]->id.'">';
+                    if(!$is_confirmed)
+                    {
+                        echo $form->field($submitteds[$j], "[$j]value")->label('Балл');
+                        echo  $form->field($submitteds[$j], 'criteria_id')->hiddenInput(['value'=> $criterias[$i]->id])->label(false);
+                        echo '<input type="hidden" name="id" value="'.$submitteds[$j]->id.'">';
+                    }
+                    else
+                    {
+                        echo $val;
+                    }
                     echo '<td>';
-                    break;
                 }   
             }
             echo '<td>';
@@ -42,9 +48,11 @@ foreach($blocks as $block){
     }
 }
 echo '</table>';
-echo '<div class="form-group" style="float:right;">';
-echo Html::submitButton('Подтвердить', ['class' => 'btn btn-success']);
-echo '</div>';
+if(!$is_confirmed){
+    echo '<div class="form-group" style="float:right;">';
+    echo Html::submitButton('Подтвердить', ['class' => 'btn btn-success']);
+    echo '</div>';
+}
 echo '</body>';
 ActiveForm::end();
 ?>
